+++ eleventy.config.js
@@ -176,7 +176,7 @@ export default async function (eleventyConfig) {
   eleventyConfig.addWatchTarget("src/assets/css/");
   eleventyConfig.addWatchTarget("src/assets/js/");
   eleventyConfig.addWatchTarget("src/blocks/");
-  eleventyConfig.addWatchTarget("content/guidelines/writing.md");
+  eleventyConfig.addWatchTarget("content/guidelines/");
 
   eleventyConfig.addPassthroughCopy({
     CNAME: "CNAME",
@@ -186,7 +186,7 @@ export default async function (eleventyConfig) {
     "src/assets/js": "assets/js",
     "node_modules/alpinejs/dist/cdn.min.js": "assets/js/alpine.min.js",
     "node_modules/@tailwindplus/elements/dist/index.js": "assets/js/elements.js",
-    "content/guidelines/writing.md": "guidelines/writing.md",
+    "content/guidelines": "guidelines",
   });
 
   eleventyConfig.addFilter("readableDate", (date) =>